refactor: 5단계 자동차 경주(리팩토링)
